+++ src/network/peernet.h
@@ -33,7 +33,7 @@ public:
     IDelegatedChannel()
       : IIOModule("delegatedchannel") {}
     virtual void PrimaryUpdate(int nStartHeight,
-                               const std::vector<std::pair<int, std::map<CDestination, size_t>>>& vEnrolledWeight,
+                               const std::vector<std::pair<uint256, std::map<CDestination, size_t>>>& vEnrolledWeight,
                                const std::map<CDestination, std::vector<unsigned char>>& mapDistributeData,
                                const std::map<CDestination, std::vector<unsigned char>>& mapPublishData)
         = 0;
